requirements.txt
requests==2.32.3
PyYAML==6.0.2
Jinja2==3.1.4
feedparser==6.0.11
python-dateutil==2.9.0.post0
